fix(auth): 修复登录重定向后页面未刷新的问题
- 在 token 过期跳转登录时，添加页面强制刷新保证状态同步
- 增强刷新令牌失败时的异常处理，确保页面跳转且刷新

refactor(chat): 优化消息ID数据类型为 Integer

- 将 ChatMessage、ChatMessageVO 等实体类的 id 和 parentId 类型由 Long 改为 Integer
- 修改 ChatSession 中 currentMessageId 类型为 Integer
- 统一更新相关服务接口及实现类对应参数类型
- 调整控制器参数类型，确保前后端类型一致

feat(chat): 在用户消息后添加助手响应加载动画

- 当最后一条消息角色为 user 时，自动插入一个正在流式加载的 assistant 响应占位消息
- 优化聊天界面消息流式呈现体验

+++ agent/src/main/java/com/hxh/apboa/agent/controller/ChatSessionController.java
@@ -52,7 +52,7 @@ public class ChatSessionController {
      * 切换历史分支（仅更新 current_message_id）
      */
     @PutMapping("/{sessionId}/current")
-    public R<Void> switchCurrentMessage(@PathVariable("sessionId") Long sessionId, @RequestParam("messageId") Long messageId) {
+    public R<Void> switchCurrentMessage(@PathVariable("sessionId") Long sessionId, @RequestParam("messageId") Integer messageId) {
         chatSessionService.switchCurrentMessage(sessionId, messageId);
         return R.success("操作成功");
     }
